development for encapsulation and baseclass and standards
