+++ Selenium4/pom.xml
@@ -44,8 +44,13 @@
 			<version>7.3.0</version>
 			<scope>test</scope>
 		</dependency>
+        <dependency>
+            <groupId>org.testng</groupId>
+            <artifactId>testng</artifactId>
+            <version>7.3.0</version>
+            <scope>compile</scope>
+        </dependency>
 
 
-
-	</dependencies>
+    </dependencies>
 </project>
