+++ Selenium4/src/main/java/Resources/datafile.properties
@@ -1,0 +1,11 @@
+herokuAppURL = https://admin:admin@the-internet.herokuapp.com/digest_auth
+herokuAppSuccessMessage = Congratulations! You must have the proper credentials.
+
+rahulshettyacademyURL = https://www.rahulshettyacademy.com/AutomationPractice/
+rahulshettyCountry = Ind
+
+dragDropURL = https://demoqa.com/droppable
+droppedText = Dropped!
+
+formURL = https://www.seleniumeasy.com/test/input-form-demo.html
+firstName = Jigar
