+++ Selenium4/src/main/java/pages/HerokuAppPage.java
@@ -1,0 +1,23 @@
+package pages;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.PageFactory;
+
+public class HerokuAppPage {
+
+    private WebDriver driver;
+
+    public HerokuAppPage(WebDriver driver) {
+        PageFactory.initElements(driver, this);
+        this.driver = driver;
+    }
+
+    @FindBy(css = "#content > div > p")
+    private WebElement successMessage;
+
+    public String getSuccessMessageText() {
+        return successMessage.getText();
+    }
+}
