+++ Selenium4/test-output/Default suite/Default test.html
@@ -75,25 +75,25 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='Selenium4.Selenium4.testSuit.scenario1()'><b>scenario1</b><br>Test class: Selenium4.Selenium4.testSuit</td>
+<td title='Selenium4.Selenium4.TestSuite.scenario1()'><b>scenario1</b><br>Test class: Selenium4.Selenium4.TestSuite</td>
 <td></td>
 <td>5</td>
-<td>Selenium4.Selenium4.testSuit@598067a5</td></tr>
+<td>Selenium4.Selenium4.TestSuite@598067a5</td></tr>
 <tr>
-<td title='Selenium4.Selenium4.testSuit.scenario2()'><b>scenario2</b><br>Test class: Selenium4.Selenium4.testSuit</td>
+<td title='Selenium4.Selenium4.TestSuite.scenario2()'><b>scenario2</b><br>Test class: Selenium4.Selenium4.TestSuite</td>
 <td></td>
 <td>21</td>
-<td>Selenium4.Selenium4.testSuit@598067a5</td></tr>
+<td>Selenium4.Selenium4.TestSuite@598067a5</td></tr>
 <tr>
-<td title='Selenium4.Selenium4.testSuit.scenario3()'><b>scenario3</b><br>Test class: Selenium4.Selenium4.testSuit</td>
+<td title='Selenium4.Selenium4.TestSuite.scenario3()'><b>scenario3</b><br>Test class: Selenium4.Selenium4.TestSuite</td>
 <td></td>
 <td>6</td>
-<td>Selenium4.Selenium4.testSuit@598067a5</td></tr>
+<td>Selenium4.Selenium4.TestSuite@598067a5</td></tr>
 <tr>
-<td title='Selenium4.Selenium4.testSuit.scenario4()'><b>scenario4</b><br>Test class: Selenium4.Selenium4.testSuit</td>
+<td title='Selenium4.Selenium4.TestSuite.scenario4()'><b>scenario4</b><br>Test class: Selenium4.Selenium4.TestSuite</td>
 <td></td>
 <td>5</td>
-<td>Selenium4.Selenium4.testSuit@598067a5</td></tr>
+<td>Selenium4.Selenium4.TestSuite@598067a5</td></tr>
 </table><p>
 </body>
 </html>
