+++ Selenium4/test-output/Default suite/Default test.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
 <testsuite hostname="PMCLAP1257" ignored="0" name="Default test" tests="4" failures="0" timestamp="26 Nov 2020 17:45:00 GMT" time="46.647" errors="0">
-  <testcase name="scenario1" time="5.354" classname="Selenium4.Selenium4.testSuit"/>
-  <testcase name="scenario2" time="21.514" classname="Selenium4.Selenium4.testSuit"/>
-  <testcase name="scenario3" time="6.134" classname="Selenium4.Selenium4.testSuit"/>
-  <testcase name="scenario4" time="5.072" classname="Selenium4.Selenium4.testSuit"/>
+  <testcase name="scenario1" time="5.354" classname="Selenium4.Selenium4.TestSuite"/>
+  <testcase name="scenario2" time="21.514" classname="Selenium4.Selenium4.TestSuite"/>
+  <testcase name="scenario3" time="6.134" classname="Selenium4.Selenium4.TestSuite"/>
+  <testcase name="scenario4" time="5.072" classname="Selenium4.Selenium4.TestSuite"/>
 </testsuite> <!-- Default test -->
