+++ Selenium4/test-output/emailable-report.html
@@ -11,11 +11,11 @@
 <tr><th colspan="7">Default suite</th></tr>
 <tr><td><a href="#t0">Default test</a></td><td class="num">4</td><td class="num">0</td><td class="num">0</td><td class="num">46,647</td><td></td><td></td></tr>
 </table>
-<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="4">Selenium4.Selenium4.testSuit</td><td><a href="#m0">scenario1</a></td><td rowspan="1">1606412661482</td><td rowspan="1">5354</td></tr><tr class="passedeven"><td><a href="#m1">scenario2</a></td><td rowspan="1">1606412666839</td><td rowspan="1">21514</td></tr><tr class="passedeven"><td><a href="#m2">scenario3</a></td><td rowspan="1">1606412688360</td><td rowspan="1">6134</td></tr><tr class="passedeven"><td><a href="#m3">scenario4</a></td><td rowspan="1">1606412694497</td><td rowspan="1">5072</td></tr></tbody>
+<table id='summary'><thead><tr><th>Class</th><th>Method</th><th>Start</th><th>Time (ms)</th></tr></thead><tbody><tr><th colspan="4">Default suite</th></tr></tbody><tbody id="t0"><tr><th colspan="4">Default test &#8212; passed</th></tr><tr class="passedeven"><td rowspan="4">Selenium4.Selenium4.TestSuite</td><td><a href="#m0">scenario1</a></td><td rowspan="1">1606412661482</td><td rowspan="1">5354</td></tr><tr class="passedeven"><td><a href="#m1">scenario2</a></td><td rowspan="1">1606412666839</td><td rowspan="1">21514</td></tr><tr class="passedeven"><td><a href="#m2">scenario3</a></td><td rowspan="1">1606412688360</td><td rowspan="1">6134</td></tr><tr class="passedeven"><td><a href="#m3">scenario4</a></td><td rowspan="1">1606412694497</td><td rowspan="1">5072</td></tr></tbody>
 </table>
-<h2>Default test</h2><h3 id="m0">Selenium4.Selenium4.testSuit#scenario1</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m1">Selenium4.Selenium4.testSuit#scenario2</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m2">Selenium4.Selenium4.testSuit#scenario3</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
-<h3 id="m3">Selenium4.Selenium4.testSuit#scenario4</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h2>Default test</h2><h3 id="m0">Selenium4.Selenium4.TestSuite#scenario1</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m1">Selenium4.Selenium4.TestSuite#scenario2</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m2">Selenium4.Selenium4.TestSuite#scenario3</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
+<h3 id="m3">Selenium4.Selenium4.TestSuite#scenario4</h3><table class="result"><tr><th class="invisible"/></tr></table><p class="totop"><a href="#summary">back to summary</a></p>
 </body>
 </html>
