+++ Selenium4/test-output/index.html
@@ -102,22 +102,22 @@
                     <div class="method-list-content passed suite-Default_suite">
                       <span>
                         <img width="3%" src="passed.png"/>
-                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="Selenium4.Selenium4.testSuit" hash-for-method="scenario1">scenario1</a> <!-- method navigator-link -->
+                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="Selenium4.Selenium4.TestSuite" hash-for-method="scenario1">scenario1</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                       <span>
                         <img width="3%" src="passed.png"/>
-                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="Selenium4.Selenium4.testSuit" hash-for-method="scenario2">scenario2</a> <!-- method navigator-link -->
+                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="Selenium4.Selenium4.TestSuite" hash-for-method="scenario2">scenario2</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                       <span>
                         <img width="3%" src="passed.png"/>
-                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="Selenium4.Selenium4.testSuit" hash-for-method="scenario3">scenario3</a> <!-- method navigator-link -->
+                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="Selenium4.Selenium4.TestSuite" hash-for-method="scenario3">scenario3</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                       <span>
                         <img width="3%" src="passed.png"/>
-                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="Selenium4.Selenium4.testSuit" hash-for-method="scenario4">scenario4</a> <!-- method navigator-link -->
+                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="Selenium4.Selenium4.TestSuite" hash-for-method="scenario4">scenario4</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                     </div> <!-- method-list-content passed suite-Default_suite -->
@@ -135,7 +135,7 @@
           <div class="suite-Default_suite-class-passed">
             <div class="main-panel-header rounded-window-top">
               <img src="passed.png"/>
-              <span class="class-name">Selenium4.Selenium4.testSuit</span>
+              <span class="class-name">Selenium4.Selenium4.TestSuite</span>
             </div> <!-- main-panel-header rounded-window-top -->
             <div class="main-panel-content rounded-window-bottom">
               <div class="method">
@@ -180,7 +180,7 @@
 &lt;suite guice-stage=&quot;DEVELOPMENT&quot; name=&quot;Default suite&quot;&gt;
   &lt;test thread-count=&quot;5&quot; verbose=&quot;2&quot; name=&quot;Default test&quot;&gt;
     &lt;classes&gt;
-      &lt;class name=&quot;Selenium4.Selenium4.testSuit&quot;/&gt;
+      &lt;class name=&quot;Selenium4.Selenium4.TestSuite&quot;/&gt;
     &lt;/classes&gt;
   &lt;/test&gt; &lt;!-- Default test --&gt;
 &lt;/suite&gt; &lt;!-- Default suite --&gt;
@@ -223,19 +223,19 @@ data.addColumn('number', 'Time (ms)');
 data.addRows(4);
 data.setCell(0, 0, 0)
 data.setCell(0, 1, 'scenario2')
-data.setCell(0, 2, 'Selenium4.Selenium4.testSuit')
+data.setCell(0, 2, 'Selenium4.Selenium4.TestSuite')
 data.setCell(0, 3, 21514);
 data.setCell(1, 0, 1)
 data.setCell(1, 1, 'scenario3')
-data.setCell(1, 2, 'Selenium4.Selenium4.testSuit')
+data.setCell(1, 2, 'Selenium4.Selenium4.TestSuite')
 data.setCell(1, 3, 6134);
 data.setCell(2, 0, 2)
 data.setCell(2, 1, 'scenario1')
-data.setCell(2, 2, 'Selenium4.Selenium4.testSuit')
+data.setCell(2, 2, 'Selenium4.Selenium4.TestSuite')
 data.setCell(2, 3, 5354);
 data.setCell(3, 0, 3)
 data.setCell(3, 1, 'scenario4')
-data.setCell(3, 2, 'Selenium4.Selenium4.testSuit')
+data.setCell(3, 2, 'Selenium4.Selenium4.TestSuite')
 data.setCell(3, 3, 5072);
 window.suiteTableData['Default_suite']= { tableData: data, tableDiv: 'times-div-Default_suite'}
 return data;
@@ -267,7 +267,7 @@ return data;
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <div class="chronological-class">
-              <div class="chronological-class-name">Selenium4.Selenium4.testSuit</div> <!-- chronological-class-name -->
+              <div class="chronological-class-name">Selenium4.Selenium4.TestSuite</div> <!-- chronological-class-name -->
               <div class="configuration-test before">
                 <span class="method-name">setup</span>
                 <span class="method-start">0 ms</span>
