+++ Selenium4/test-output/junitreports/TEST-Selenium4.Selenium4.testSuit.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite skipped="0" hostname="PMCLAP1257" name="Selenium4.Selenium4.testSuit" tests="4" failures="0" timestamp="26 Nov 2020 17:45:00 GMT" time="38.074" errors="0">
-  <testcase name="scenario2" time="21.514" classname="Selenium4.Selenium4.testSuit"/>
-  <testcase name="scenario1" time="5.354" classname="Selenium4.Selenium4.testSuit"/>
-  <testcase name="scenario4" time="5.072" classname="Selenium4.Selenium4.testSuit"/>
-  <testcase name="scenario3" time="6.134" classname="Selenium4.Selenium4.testSuit"/>
-</testsuite> <!-- Selenium4.Selenium4.testSuit -->
+<testsuite skipped="0" hostname="PMCLAP1257" name="Selenium4.Selenium4.TestSuite" tests="4" failures="0" timestamp="26 Nov 2020 17:45:00 GMT" time="38.074" errors="0">
+  <testcase name="scenario2" time="21.514" classname="Selenium4.Selenium4.TestSuite"/>
+  <testcase name="scenario1" time="5.354" classname="Selenium4.Selenium4.TestSuite"/>
+  <testcase name="scenario4" time="5.072" classname="Selenium4.Selenium4.TestSuite"/>
+  <testcase name="scenario3" time="6.134" classname="Selenium4.Selenium4.TestSuite"/>
+</testsuite> <!-- Selenium4.Selenium4.TestSuite -->
