+++ Selenium4/test-output/old/Default suite/classes.html
@@ -4,7 +4,7 @@
 <th>Method name</th>
 <th>Groups</th>
 </tr><tr>
-<td>Selenium4.Selenium4.testSuit</td>
+<td>Selenium4.Selenium4.TestSuite</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@Test</td>
