+++ Selenium4/test-output/old/Default suite/methods-alphabetical.html
@@ -1,16 +1,16 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:21</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="testSuit.scenario1()[pri:0, instance:Selenium4.Selenium4.testSuit@598067a5]">scenario1</td> 
+<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:21</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="testSuit.scenario1()[pri:0, instance:Selenium4.Selenium4.TestSuite@598067a5]">scenario1</td>
   <td>main@174573182</td>   <td></td> </tr>
-<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:26</td>   <td>5357</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="testSuit.scenario2()[pri:0, instance:Selenium4.Selenium4.testSuit@598067a5]">scenario2</td> 
+<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:26</td>   <td>5357</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="testSuit.scenario2()[pri:0, instance:Selenium4.Selenium4.TestSuite@598067a5]">scenario2</td>
   <td>main@174573182</td>   <td></td> </tr>
-<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:48</td>   <td>26878</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="testSuit.scenario3()[pri:0, instance:Selenium4.Selenium4.testSuit@598067a5]">scenario3</td> 
+<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:48</td>   <td>26878</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="testSuit.scenario3()[pri:0, instance:Selenium4.Selenium4.TestSuite@598067a5]">scenario3</td>
   <td>main@174573182</td>   <td></td> </tr>
-<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:54</td>   <td>33015</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="testSuit.scenario4()[pri:0, instance:Selenium4.Selenium4.testSuit@598067a5]">scenario4</td> 
+<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:54</td>   <td>33015</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="testSuit.scenario4()[pri:0, instance:Selenium4.Selenium4.TestSuite@598067a5]">scenario4</td>
   <td>main@174573182</td>   <td></td> </tr>
-<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:13</td>   <td>-7778</td> <td>&nbsp;</td><td title="&gt;&gt;testSuit.setup()[pri:0, instance:Selenium4.Selenium4.testSuit@598067a5]">&gt;&gt;setup</td> 
+<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:13</td>   <td>-7778</td> <td>&nbsp;</td><td title="&gt;&gt;testSuit.setup()[pri:0, instance:Selenium4.Selenium4.TestSuite@598067a5]">&gt;&gt;setup</td>
 <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@174573182</td>   <td></td> </tr>
-<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:59</td>   <td>38091</td> <td>&nbsp;</td><td title="&lt;&lt;testSuit.tearDown()[pri:0, instance:Selenium4.Selenium4.testSuit@598067a5]">&lt;&lt;tearDown</td> 
+<tr bgcolor="f5e5e3">  <td>20/11/26 23:14:59</td>   <td>38091</td> <td>&nbsp;</td><td title="&lt;&lt;testSuit.tearDown()[pri:0, instance:Selenium4.Selenium4.TestSuite@598067a5]">&lt;&lt;tearDown</td>
 <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@174573182</td>   <td></td> </tr>
 </table>
